+++ src/pages/membership/helpers/__tests__/createProfileMutationVariables.test.ts
@@ -1,4 +1,4 @@
-import { FormValues } from '../../components/createYouthProfileForm/CreateYouthProfileForm';
+import { FormValues } from '../../components/youthProfileForm/YouthProfileForm';
 import {
   PrefillRegistartion,
   YouthLanguage,
